requirements.txt
Django>=4.2
djangorestframework>=3.14
pytest
pytest-django
flake8
psycopg2-binary
dj_database_url


